fix small issue

+++ tests/kernels/test_fake_quantize.cpp
@@ -56,7 +56,7 @@ TEST_P(FakeQuantizeTest, fake_quantize) {
     int8_t zero_point[] = {127};
     auto zero_point_ptr =
         hrt::create(nncase::dt_int8, {1},
-                    {reinterpret_cast<gsl::byte *>(zero_point), sizeof(float)},
+                    {reinterpret_cast<gsl::byte *>(zero_point), sizeof(int8_t)},
                     true, host_runtime_tensor::pool_cpu_only)
             .expect("create tensor failed");
